Firmware: EEPROMWriter class updated, some minor fixes (found with clang static analisys)

+++ Firmware/OTPManagerFirmware/eeprom_writer.h
@@ -15,7 +15,7 @@ class EEPROMWriter
 	public:
 		EEPROMWriter(const int baseAddr, const int maxLen, Cipher &cipher, const uint8_t * const encKey, uint8_t * const tweak);
 		bool WriteNextBlock(const uint8_t * const data);
-		bool WriteData(const uint8_t* const data, const size_t dLen);
+		bool WriteData(const uint8_t* const data, const uint16_t dLen);
 };
 
 #endif
